Infra/modules/account-sync/function_source/requirements.txt
functions-framework==3.*
google-api-python-client==2.*
google-auth==2.*
google-cloud-storage==2.*
google-cloud-logging